requirements.txt
ultralytics>=8.0.0
fast-alpr[onnxruntime]
onnxruntime
opencv-python
numpy
pandas
streamlit
supervision
Pillow